# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cellier API",
...
    "contact": {
      "name": "Equipo Cellier"
    },
    "license": {
      "name": "Uso interno"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/households/{householdId}": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Ver un hogar",
        "description": "Detalle del hogar. Lo puede consultar cualquier miembro, con rol\n`ADMIN` o `MEMBER`.\n\nEs el **único** punto de la API por el que sale el `joinCode`, y solo\ndespués de comprobar la membresía: cualquier miembro puede compartir el\ncódigo, pero nadie ajeno al hogar llega a verlo.\n",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del hogar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetail"
                },
                "examples": {
                  "detalle": {
                    "description": "detalle",
                    "value": {
                      "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                      "name": "Casa Rivas",
                      "joinCode": "K7M2QP9X",
                      "role": "ADMIN",
                      "memberCount": 3,
                      "createdAt": "2026-09-03T18:42:11Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Hogares"
        ],
        "summary": "Eliminar un hogar",
        "description": "Borra el hogar **y todo su contenido**: membresías, solicitudes de ingreso\ny, a medida que existan, despensa, plantillas y recetas. La operación no\ntiene vuelta atrás y no exige que el hogar se quede antes sin miembros.\n\nSolo un administrador puede hacerlo.\n",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          }
        ],
        "responses": {
          "204": {
            "description": "Hogar eliminado."
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "El usuario es miembro del hogar, pero no administrador.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "noEsAdmin": {
                    "description": "noEsAdmin",
                    "value": {
                      "type": "https://cellier.app/problems/forbidden",
                      "title": "Operación no permitida",
                      "status": 403,
                      "detail": "Esta operación requiere rol de administrador en el hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Hogares"
        ],
        "summary": "Renombrar un hogar",
        "description": "Cambia el nombre del hogar. Solo un administrador puede hacerlo.",
        "operationId": "rename",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Hogar actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetail"
                },
                "examples": {
                  "detalle": {
                    "description": "detalle",
                    "value": {
                      "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                      "name": "Casa Rivas",
                      "joinCode": "K7M2QP9X",
                      "role": "ADMIN",
                      "memberCount": 3,
                      "createdAt": "2026-09-03T18:42:11Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El nombre falta, está en blanco o pasa de 80 caracteres.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "nombreVacio": {
                    "description": "nombreVacio",
                    "value": {
                      "type": "https://cellier.app/problems/bad-request",
                      "title": "Datos inválidos",
                      "status": 400,
                      "detail": "Uno o más campos del cuerpo de la petición no son válidos.",
                      "instance": "/api/v1/households",
                      "errors": {
                        "name": "El nombre del hogar es obligatorio"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "El usuario es miembro del hogar, pero no administrador.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "noEsAdmin": {
                    "description": "noEsAdmin",
                    "value": {
                      "type": "https://cellier.app/problems/forbidden",
                      "title": "Operación no permitida",
                      "status": 403,
                      "detail": "Esta operación requiere rol de administrador en el hogar.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/households/{householdId}/members": {
      "get": {
        "tags": [
          "Hogares"
        ],
        "summary": "Listar los miembros del hogar",
        "description": "Las personas del hogar con su rol y su fecha de ingreso. Lo puede consultar\ncualquier miembro, sea `ADMIN` o `MEMBER`.\n\nLlegan primero los administradores y, dentro de cada grupo, por antigüedad.\n",
        "operationId": "list_3",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "description": "Identificador del hogar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          }
        ],
        "responses": {
          "200": {
            "description": "Miembros del hogar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdMember"
                  }
                },
                "examples": {
                  "miembros": {
                    "description": "miembros",
                    "value": [
                      {
                        "userId": "3f1c9a2e-5b7d-4a10-9c88-2f0e6b4d1a73",
                        "displayName": "Ana Rivas",
                        "email": "ana.rivas@gmail.com",
                        "avatarUrl": "https://lh3.googleusercontent.com/a/ACg8ocK",
                        "role": "ADMIN",
                        "joinedAt": "2026-09-03T18:42:11Z"
                      },
                      {
                        "userId": "d90f4c21-77ab-4e35-9f62-1c8b03ad5e47",
                        "displayName": "Bruno Soto",
                        "email": "bruno.soto@gmail.com",
                        "avatarUrl": null,
                        "role": "MEMBER",
                        "joinedAt": "2026-09-04T09:15:02Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falta el access token, o no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "sinToken": {
                    "description": "sinToken",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "Se requiere un access token válido.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/members"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "El hogar no existe, o existe pero el usuario autenticado no es miembro. La API no distingue ambos casos a propósito.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "hogarAjenoOInexistente": {
                    "description": "hogarAjenoOInexistente",
                    "value": {
                      "type": "https://cellier.app/problems/not-found",
                      "title": "Recurso no encontrado",
                      "status": 404,
                      "detail": "No existe ese hogar, o no eres miembro de él.",
                      "instance": "/api/v1/households/8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98/members"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "HouseholdDetail": {
        "type": "object",
        "description": "Detalle de un hogar, incluido su código de ingreso.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del hogar.",
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas"
          },
          "joinCode": {
            "type": "string",
            "description": "Código para solicitar el ingreso. Conocerlo no da acceso: abre una solicitud\nque un administrador debe aprobar.",
            "example": "K7M2QP9X"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario autenticado en este hogar.",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "description": "Número de personas que pertenecen al hogar.",
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del hogar.",
            "example": "2026-09-03T18:42:11Z"
          }
        }
      },
...
      "UpdateHouseholdRequest": {
        "type": "object",
        "description": "Cambios a aplicar sobre un hogar.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nuevo nombre del hogar.",
            "example": "Casa Rivas Soto",
            "maxLength": 80,
            "minLength": 0
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT emitido por Cellier.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}